requirements.txt
fastapi
uvicorn
pydantic
sentence-transformers
faiss-cpu
google-generativeai
pandas
numpy
